moving component components/home/Search to its own folder + styled components
